Improve new command git interactions
The GITHUB_ACCOUNT parameter name was change to GITHUB_OWNER to
match the new repository page on Github.  Also, GITHUB_ACCOUNT
was confusing as to whether it should be the current user's account
or the owners account.

A check was added to insure that the current user has SSH setup
with Github.  If they do not then the new command aborts and provides
instructions on how to setup SSH with github.