Add --slurm option to md5 command
If md5 is ran against large raw data files it can take a long time
and it can overload login nodes.  This option is intended to fix
these two issues.

The --slurm option has the valid values "batch" and "run".  If the
value is batch then it runs as a Slurm sbatch job with the output
written to a file named lab-run-<job_id>.out.  If the value is run
then it runs as a Slurm job connected to the current terminal
output using srun.