Add git related options to lab new
The --skip-git option creates a project without making it a git
repository so other source control tools can be used.

The -o, --owner options allows the git remote owner to be specified
so that git remotes can be added for hosting providers like github
and gitlanb.

The --git-host option allows the remotes hosting provider to be
specified for creating git remotes.  The default is github.com.